backend/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
pyyaml>=6.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
openai>=1.0.0
# Web search for AI product finder (set TAVILY_API_KEY)
tavily-python>=0.5.0
requests>=2.32.0
